fix : 필터링 에러 수정
