+++ public/index.html
@@ -10,8 +10,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <!-- mixed 에러 해결 -->
-    <meta http-equiv="Content-Security-Policy" content="upgrade-insecure-requests" />
+
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
