+++ src/hooks/useFilteringMarker.ts
@@ -4,6 +4,7 @@ import axios from 'axios';
 import filterState from '../recoil/atoms';
 import { generateMarker } from '../utils/mapUtils';
 import { FacilitiesType } from '../types/mapTypes';
+import useToast from './useToast';
 
 interface UseFilteringMarkerProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -33,44 +34,53 @@ function addMarkers(map: any, key: string, positions: FilteringPosition[]) {
   const newMarkers = positions.map(position =>
     generateMarker(map, position.lat, position.lon, position.type)
   );
-  newMarkers.forEach(marker => marker.setMap(map));
   return { key, markers: newMarkers };
 }
 
 function useFilteringMarker({ map, lat, lng }: UseFilteringMarkerProps): void {
   const filterValue = useRecoilValue(filterState);
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const [position, setPosition] = useState<{[key : string] : FilteringPosition[]} | undefined>({});
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const [markers, setMarkers] = useState<{ [key: string]: any[] }>({});
+  const { createToast } = useToast();
 
   useEffect(() => {
-    if (!lat || !lng) return;
+    if(!lat || !lng) return;
 
-    const trueKeys = Object.keys(filterValue).filter(key => filterValue[key]);
-    if (!trueKeys.length) {
-      // 필터된 값에서 true인 값이 하나도 없으면 모든 마커 제거
-      Object.values(markers).flat().forEach(marker => marker.setMap(null));
-      setMarkers({});
-      return;
-    }
+    const allFilter = ['firestation', 'safetycenter', 'cctv', 'safetyfacility', 'heatshelter', 'emergencybell'];
 
-    Promise.all(trueKeys.map(key => fetchPositionData(key, lat, lng)))
+    Promise.all(allFilter.map(key => fetchPositionData(key, lat, lng)))
       .then(responses => {
-        // 이전 마커 제거
-        Object.values(markers).flat().forEach(marker => marker.setMap(null));
-
-        // 새로운 마커 추가
-        const newMarkers = responses.reduce((acc, res) => {
-          const { key, markers: markersArray } = addMarkers(map, res.key, res.data);
-          return { ...acc, [key]: markersArray };
-        }, {});
-
-        // 새로운 마커 설정
-        setMarkers(newMarkers);
+        setPosition(responses.reduce((acc, res) => {
+          return { ...acc, [res.key]: res.data };
+        }, {}));
       })
       .catch(err => {
+        createToast({
+          msg: '마커 데이터를 불러오는데 실패했습니다.',
+          type: 'error'
+        });
         console.error(err);
       });
-  }, [filterValue, lat, lng, map]);
+  }, []);
+
+  useEffect(() => {
+    // 이상해...
+    if(!position) return;
+    const trueKeys = Object.keys(filterValue).filter(key => filterValue[key]);
+    console.log('trueKeys는 이거야!!',trueKeys);
+    if(markers) {
+      Object.values(markers).flat().forEach(marker => marker.setMap(null));
+      console.log('이전 마커들을 flat 한거', Object.values(markers).flat());
+      setMarkers({});
+    }
+    trueKeys.forEach(key => {
+      const { markers: markersArray } = addMarkers(map, key, position[key]);
+      setMarkers((prev) => ({ ...prev, [key]: markersArray }));
+    });
+
+  }, [map, filterValue]);
 }
 
 export default useFilteringMarker;
